+++ roles/scheduler/scheduler.go
@@ -3,6 +3,7 @@ package scheduler
 import (
 	"encoding/json"
 	"errors"
+	"fmt"
 	"strings"
 	"time"
 
@@ -10,6 +11,7 @@ import (
 	"github.com/nadavbm/etzba/pkg/reader"
 	"github.com/nadavbm/etzba/roles/apiclient"
 	"github.com/nadavbm/etzba/roles/authenticator"
+	"github.com/nadavbm/etzba/roles/sqlclient"
 	"github.com/nadavbm/etzba/roles/worker"
 	"github.com/nadavbm/zlog"
 	"go.uber.org/zap"
@@ -212,6 +214,35 @@ func (s *Scheduler) executeAPIRequestFromAssignment(assigment *worker.Assignment
 	}
 	return worker.GetAPIRequestDuration(assigment)
 
+}
+
+// prepareAssignmentsForResultCollection set maps to prepare the result output with assignment durations and responses
+func (s *Scheduler) prepareAssignmentsForResultCollection(assignments []worker.Assignment) (map[string][]time.Duration, map[string][]*apiclient.Response, error) {
+	allAssignmentsExecutionsDurations := make(map[string][]time.Duration)
+	allAssignmentsExecutionsResponses := make(map[string][]*apiclient.Response)
+	var allAPIResponses []*apiclient.Response
+	var allDurations []time.Duration
+	for _, a := range assignments {
+		allAssignmentsExecutionsDurations[getAssignmentAsString(a, s.ExecutionType)] = allDurations
+		allAssignmentsExecutionsResponses[getAssignmentAsString(a, s.ExecutionType)] = allAPIResponses
+	}
+
+	return allAssignmentsExecutionsDurations, allAssignmentsExecutionsResponses, nil
+}
+
+//
+// ------------------------------------------------------------------------------- helpers ------------------------------------------------------------------
+//
+
+// getAssignmentAsString prepare the assignment title for stdout
+func getAssignmentAsString(a worker.Assignment, command string) string {
+	switch {
+	case command == "api":
+		return fmt.Sprintf("URL: %s, Method: %s", a.ApiRequest.Url, a.ApiRequest.Method)
+	case command == "sql":
+		return fmt.Sprintf("%s", sqlclient.ToSQL(&a.SqlQuery))
+	}
+	return ""
 }
 
 /*
